Set up integration tests (#4)
A simple server for the Lunaria API has been implemented to enable integration
tests. The server returns static content that can be used for assertions in the
client libraries.

Right now, the server has to be run in the background when running tests. The
recommended way to run the tests is through the scripts in bin, as they start
the server in a background process before starting the test suite.