Create Node package (#5)
As a second proof-of-concept, a Node client has been generated and added
to the project. The second package was added for two reasons:

  1. To enable the development of a release workflow that publishes all
     libraries at once
  2. To create an example where the Protocol Buffers are compiled to
     language-specific code

Node has been chosen since it will probably be used the most by players
to create their bots for Lunaria. The auto-generated source files are
included in the repository for convenience, but get recreated during
continuous integration to always be up-to-date when testing and
publishing the package.