Symlink Protobufs into Rust crate
Rust requires the Protocol Buffers to be distributed with the crate,
because otherwise the code generation will fail. This has not shown up
as an issue yet because we used relative paths for the dependency so
far, but now that the crate has been published to crates.io it got
reported.

We still want to have the Protocol Buffers at the root of the
repository, because they are not a feature specific to the Rust
implementation. So they have been symlinked into the Rust crate.
Symlinks can cause some issues with Git, so this is considered a
temporary workaround until a proper solution can be found.

See <https://github.com/playlunaria/lunaria/issues/20> for the bug
report.