Run checks on each push
We missed a regression in the last commit, since the checks for Node
didn't run based on our configuration. We only changed the Protocol
Buffers, which had an impact on the Node module. But because no
Javascript or TypeScript file changed, the Node action wasn't run and
the regression was missed.

While we could run the Node action when the Protocol Buffers change as
well, it is safer to simly default to running all actions al the time.
We don't expect this repository to change very frequently, and the
current actions are fast enough as they are.